migrate more layout stuff to `data`
