+++ data/src/lib.rs
@@ -1,3 +1,228 @@
+use std::fmt;
+
+use chrono::DateTime;
+use exchanges::{Ticker, adapter::Exchange};
+use iced_core::{Point, Size, Theme};
+use pane::Pane;
+use serde::{Deserialize, Serialize};
+use theme::SerializableTheme;
+
 pub mod charts;
-pub mod layout;
 pub mod pane;
+pub mod theme;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Layout {
+    pub name: String,
+    pub dashboard: Dashboard,
+}
+
+impl Default for Layout {
+    fn default() -> Self {
+        Self {
+            name: "Default".to_string(),
+            dashboard: Dashboard::default(),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Layouts {
+    pub layouts: Vec<Layout>,
+    pub active_layout: String,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct Dashboard {
+    pub pane: Pane,
+    pub popout: Vec<(Pane, (f32, f32), (f32, f32))>,
+    pub trade_fetch_enabled: bool,
+}
+
+impl Default for Dashboard {
+    fn default() -> Self {
+        Self {
+            pane: Pane::Starter,
+            popout: vec![],
+            trade_fetch_enabled: false,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq)]
+pub struct ScaleFactor(f64);
+
+impl Default for ScaleFactor {
+    fn default() -> Self {
+        Self(1.0)
+    }
+}
+
+impl From<f64> for ScaleFactor {
+    fn from(value: f64) -> Self {
+        ScaleFactor(value.clamp(0.8, 1.8))
+    }
+}
+
+impl From<ScaleFactor> for f64 {
+    fn from(value: ScaleFactor) -> Self {
+        value.0
+    }
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Copy, Deserialize, Serialize)]
+pub enum Sidebar {
+    #[default]
+    Left,
+    Right,
+}
+
+impl std::fmt::Display for Sidebar {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Sidebar::Left => write!(f, "Left"),
+            Sidebar::Right => write!(f, "Right"),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct State {
+    pub layout_manager: Layouts,
+    pub selected_theme: SerializableTheme,
+    pub favorited_tickers: Vec<(Exchange, Ticker)>,
+    pub window_size: Option<(f32, f32)>,
+    pub window_position: Option<(f32, f32)>,
+    pub timezone: UserTimezone,
+    pub sidebar: Sidebar,
+    pub scale_factor: ScaleFactor,
+}
+
+impl State {
+    pub fn from_parts(
+        layout_manager: Layouts,
+        selected_theme: Theme,
+        favorited_tickers: Vec<(Exchange, Ticker)>,
+        size: Option<Size>,
+        position: Option<Point>,
+        timezone: UserTimezone,
+        sidebar: Sidebar,
+        scale_factor: ScaleFactor,
+    ) -> Self {
+        State {
+            layout_manager,
+            selected_theme: SerializableTheme {
+                theme: selected_theme,
+            },
+            favorited_tickers,
+            window_size: size.map(|s| (s.width, s.height)),
+            window_position: position.map(|p| (p.x, p.y)),
+            timezone,
+            sidebar,
+            scale_factor,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Default)]
+pub enum UserTimezone {
+    #[default]
+    Utc,
+    Local,
+}
+
+impl UserTimezone {
+    /// Converts UTC timestamp to the appropriate timezone and formats it according to timeframe
+    pub fn format_timestamp(&self, timestamp: i64, timeframe: u64) -> String {
+        if let Some(datetime) = DateTime::from_timestamp(timestamp, 0) {
+            match self {
+                UserTimezone::Local => {
+                    let time_with_zone = datetime.with_timezone(&chrono::Local);
+                    Self::format_by_timeframe(time_with_zone, timeframe)
+                }
+                UserTimezone::Utc => {
+                    let time_with_zone = datetime.with_timezone(&chrono::Utc);
+                    Self::format_by_timeframe(time_with_zone, timeframe)
+                }
+            }
+        } else {
+            String::new()
+        }
+    }
+
+    /// Formats a `DateTime` with appropriate format based on timeframe
+    fn format_by_timeframe<Tz: chrono::TimeZone>(datetime: DateTime<Tz>, timeframe: u64) -> String
+    where
+        Tz::Offset: std::fmt::Display,
+    {
+        if timeframe < 10000 {
+            datetime.format("%M:%S").to_string()
+        } else if datetime.format("%H:%M").to_string() == "00:00" {
+            datetime.format("%-d").to_string()
+        } else {
+            datetime.format("%H:%M").to_string()
+        }
+    }
+
+    /// Formats a `DateTime` with detailed format for crosshair display
+    pub fn format_crosshair_timestamp(&self, timestamp_millis: i64, timeframe: u64) -> String {
+        if let Some(datetime) = DateTime::from_timestamp_millis(timestamp_millis) {
+            if timeframe < 10000 {
+                return datetime.format("%M:%S:%3f").to_string().replace('.', "");
+            }
+
+            match self {
+                UserTimezone::Local => datetime
+                    .with_timezone(&chrono::Local)
+                    .format("%a %b %-d  %H:%M")
+                    .to_string(),
+                UserTimezone::Utc => datetime
+                    .with_timezone(&chrono::Utc)
+                    .format("%a %b %-d  %H:%M")
+                    .to_string(),
+            }
+        } else {
+            String::new()
+        }
+    }
+}
+
+impl fmt::Display for UserTimezone {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            UserTimezone::Utc => write!(f, "UTC"),
+            UserTimezone::Local => {
+                let local_offset = chrono::Local::now().offset().local_minus_utc();
+                let hours = local_offset / 3600;
+                let minutes = (local_offset % 3600) / 60;
+                write!(f, "Local (UTC {hours:+03}:{minutes:02})")
+            }
+        }
+    }
+}
+
+impl<'de> Deserialize<'de> for UserTimezone {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let timezone_str = String::deserialize(deserializer)?;
+        match timezone_str.to_lowercase().as_str() {
+            "utc" => Ok(UserTimezone::Utc),
+            "local" => Ok(UserTimezone::Local),
+            _ => Err(serde::de::Error::custom("Invalid UserTimezone")),
+        }
+    }
+}
+
+impl Serialize for UserTimezone {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        match self {
+            UserTimezone::Utc => serializer.serialize_str("UTC"),
+            UserTimezone::Local => serializer.serialize_str("Local"),
+        }
+    }
+}
