+++ data/src/theme.rs
@@ -1,0 +1,102 @@
+use iced_core::{
+    Color, Theme,
+    theme::{Custom, Palette},
+};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone)]
+pub struct SerializableTheme {
+    pub theme: Theme,
+}
+
+impl Default for SerializableTheme {
+    fn default() -> Self {
+        Self {
+            theme: Theme::Custom(custom_theme().into()),
+        }
+    }
+}
+
+pub fn custom_theme() -> Custom {
+    Custom::new(
+        "Flowsurface".to_string(),
+        Palette {
+            background: Color::from_rgb8(24, 22, 22),
+            text: Color::from_rgb8(197, 201, 197),
+            primary: Color::from_rgb8(200, 200, 200),
+            success: Color::from_rgb8(81, 205, 160),
+            danger: Color::from_rgb8(192, 80, 77),
+            warning: Color::from_rgb8(238, 216, 139),
+        },
+    )
+}
+
+impl Serialize for SerializableTheme {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let theme_str = match self.theme {
+            Theme::Ferra => "ferra",
+            Theme::Dark => "dark",
+            Theme::Light => "light",
+            Theme::Dracula => "dracula",
+            Theme::Nord => "nord",
+            Theme::SolarizedLight => "solarized_light",
+            Theme::SolarizedDark => "solarized_dark",
+            Theme::GruvboxLight => "gruvbox_light",
+            Theme::GruvboxDark => "gruvbox_dark",
+            Theme::CatppuccinLatte => "catppuccino_latte",
+            Theme::CatppuccinFrappe => "catppuccino_frappe",
+            Theme::CatppuccinMacchiato => "catppuccino_macchiato",
+            Theme::CatppuccinMocha => "catppuccino_mocha",
+            Theme::TokyoNight => "tokyo_night",
+            Theme::TokyoNightStorm => "tokyo_night_storm",
+            Theme::TokyoNightLight => "tokyo_night_light",
+            Theme::KanagawaWave => "kanagawa_wave",
+            Theme::KanagawaDragon => "kanagawa_dragon",
+            Theme::KanagawaLotus => "kanagawa_lotus",
+            Theme::Moonfly => "moonfly",
+            Theme::Nightfly => "nightfly",
+            Theme::Oxocarbon => "oxocarbon",
+            Theme::Custom(_) => "flowsurface",
+        };
+        theme_str.serialize(serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for SerializableTheme {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let theme_str = String::deserialize(deserializer)?;
+        let theme = match theme_str.as_str() {
+            "ferra" => Theme::Ferra,
+            "dark" => Theme::Dark,
+            "light" => Theme::Light,
+            "dracula" => Theme::Dracula,
+            "nord" => Theme::Nord,
+            "solarized_light" => Theme::SolarizedLight,
+            "solarized_dark" => Theme::SolarizedDark,
+            "gruvbox_light" => Theme::GruvboxLight,
+            "gruvbox_dark" => Theme::GruvboxDark,
+            "catppuccino_latte" => Theme::CatppuccinLatte,
+            "catppuccino_frappe" => Theme::CatppuccinFrappe,
+            "catppuccino_macchiato" => Theme::CatppuccinMacchiato,
+            "catppuccino_mocha" => Theme::CatppuccinMocha,
+            "tokyo_night" => Theme::TokyoNight,
+            "tokyo_night_storm" => Theme::TokyoNightStorm,
+            "tokyo_night_light" => Theme::TokyoNightLight,
+            "kanagawa_wave" => Theme::KanagawaWave,
+            "kanagawa_dragon" => Theme::KanagawaDragon,
+            "kanagawa_lotus" => Theme::KanagawaLotus,
+            "moonfly" => Theme::Moonfly,
+            "nightfly" => Theme::Nightfly,
+            "oxocarbon" => Theme::Oxocarbon,
+            "flowsurface" => SerializableTheme::default().theme,
+            _ => return Err(serde::de::Error::custom("Invalid theme")),
+        };
+        Ok(SerializableTheme { theme })
+    }
+}
